requirements.txt
streamlit==1.42.0
openai==1.65.1
beautifulsoup4==4.13.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.6
